requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==12.0.0
idna==3.3
iniconfig==1.1.1
mixer==7.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
